refactor(api-gatway): wrapped exports implemented and used
