+++ API-Gateway/src/rest/index.js
@@ -1,0 +1,4 @@
+export { default as accounts } from './accounts';
+export { default as db } from './db';
+export { default as offchain } from './offchain';
+export { default as zkp } from './zkp';
