+++ API-Gateway/src/routes/index.js
@@ -1,0 +1,7 @@
+export { default as rootRouter } from './api-gateway';
+export { default as nftCommitmentRoutes } from './nft-commitment';
+export { default as ftCommitmentRoutes } from './ft-commitment';
+export { default as ftRoutes } from './ft';
+export { default as nftRoutes } from './nft';
+export { default as userRoutes } from './user';
+export { default as shieldRoutes } from './shield';
